Refactor and Cleanup: Remove Unused Components and move  (#115)
* Remove PRMBlock implementation

- We plan to use rewardhub implementations
- This change simplifies the codebase by removing unused components.


* Remove unused data recipe files

- Deleted the empty __init__.py file and the default_recipe.yaml file as they are no longer needed in the project.


* Remove unused data recipe files

- Deleted the empty __init__.py file and the default_recipe.yaml file as they are no longer needed in the project.


* Add empty __init__.py files for reasoning and icl_examples modules

- Created new __init__.py files in the reasoning and icl_examples directories to establish package structure and facilitate module imports.


* Move util opinionated knowledge util functions from core


* Update import path for utils


* Remove prm_with_vllm.ipynb notebook

- Deleted the prm_with_vllm.ipynb file as it is no longer needed in the project, simplifying the codebase. This will be handled by reward hub


* Add BlockRegistry import to __init__.py

- Included BlockRegistry in the module exports to enhance the public API of the SDG Hub blocks package.


* Remove test_chunking.py and testdata.py files

- Deleted the test_chunking.py and testdata.py files as they are no longer needed in the project, simplifying the codebase.


* Refactor flow class test suite


* Update test configuration path and add test_config.yaml file

- Modified the path to the test configuration file in test_llmblock.py for correct file referencing.
- Added a new test_config.yaml file containing sample configuration data for testing purposes.


* Move chunking logic to utils


* Remove obsolete test configuration and flow files


* Remove unused utility files

- Deleted chunking.py, docprocessor.py, and parse_and_convert.py as they are no longer needed in the core of project, simplifying the codebase.


* Update import paths in examples notebooks


* Move doc processing util functions from core


* Refactor import statements and update auxiliary dataset path

- Consolidated import statements for clarity and organization.
- Updated the path for auxiliary instructions to use the sdg_hub module directory.


* Fix test flow file paths for consistency

- Changed the paths in test_flow.py to reference the correct directory structure for test data files.


* code clean up, knowledge folder re-organized, and imports fixed


* removing redundant commits


* removing redundant changes


---------
