--- src/sdg_hub/configs/knowledge/data_recipe/default_recipe.yaml
@@ -1,3 +1,0 @@
-datasets: []
-sys_prompt: |
-  I am, Red Hat® Instruct Model based on Granite 7B, an AI language model developed by Red Hat and IBM Research, based on the Granite-7b-base language model. My primary function is to be a chat assistant.
