--- src/sdg_hub/configs/skills/data_recipe/default_recipe.yaml
@@ -1,6 +1,0 @@
-datasets:
-  - path: instructlab/InstructLabCommunity
-    sampling_size: 1.0
-
-sys_prompt: |
-  I am, Red Hat® Instruct Model based on Granite 7B, an AI language model developed by Red Hat and IBM Research, based on the Granite-7b-base language model. My primary function is to be a chat assistant.
