+++ tests/flows/test_flow.py
@@ -15,14 +15,14 @@ class TestFlow(unittest.TestCase):
         self.flow = Flow(MagicMock())
 
     def test_config_relative_to_flow(self):
-        flow = self.flow.get_flow_from_file("tests/testdata/test_flow_1.yaml")
+        flow = self.flow.get_flow_from_file("tests/flows/testdata/test_flow_1.yaml")
         block = flow[0]["block_type"](**flow[0]["block_config"])
 
         self.assertEqual(block.block_config["introduction"], "intro")
 
     def test_config_relative_to_package(self):
         with open(
-            "tests/testdata/test_flow_1.yaml", "r", encoding="utf-8"
+            "tests/flows/testdata/test_flow_1.yaml", "r", encoding="utf-8"
         ) as yaml_file:
             y = yaml.safe_load(yaml_file)
         y[0]["block_config"]["config_path"] = (
@@ -30,7 +30,7 @@ class TestFlow(unittest.TestCase):
         )
         with patch("yaml.safe_load", new_callable=MagicMock) as mock_safe_load:
             mock_safe_load.return_value = y
-            flow = self.flow.get_flow_from_file("tests/testdata/test_flow_1.yaml")
+            flow = self.flow.get_flow_from_file("tests/flows/testdata/test_flow_1.yaml")
         block = flow[0]["block_type"](**flow[0]["block_config"])
 
         self.assertEqual(
@@ -40,7 +40,7 @@ class TestFlow(unittest.TestCase):
 
     def test_config_absolute(self):
         with open(
-            "tests/testdata/test_flow_1.yaml", "r", encoding="utf-8"
+            "tests/flows/testdata/test_flow_1.yaml", "r", encoding="utf-8"
         ) as yaml_file:
             y = yaml.safe_load(yaml_file)
         y[0]["block_config"]["config_path"] = os.path.abspath(
@@ -48,7 +48,7 @@ class TestFlow(unittest.TestCase):
         )
         with patch("yaml.safe_load", new_callable=MagicMock) as mock_safe_load:
             mock_safe_load.return_value = y
-            flow = self.flow.get_flow_from_file("tests/testdata/test_flow_1.yaml")
+            flow = self.flow.get_flow_from_file("tests/flows/testdata/test_flow_1.yaml")
         block = flow[0]["block_type"](**flow[0]["block_config"])
 
         self.assertEqual(
@@ -58,7 +58,7 @@ class TestFlow(unittest.TestCase):
 
     def test_config_list_mix(self):
         with open(
-            "tests/testdata/test_flow_2.yaml", "r", encoding="utf-8"
+            "tests/flows/testdata/test_flow_2.yaml", "r", encoding="utf-8"
         ) as yaml_file:
             y = yaml.safe_load(yaml_file)
         y[0]["block_config"]["config_paths"]["k3"] = os.path.abspath(
@@ -67,7 +67,7 @@ class TestFlow(unittest.TestCase):
 
         with patch("yaml.safe_load", new_callable=MagicMock) as mock_safe_load:
             mock_safe_load.return_value = y
-            flow = self.flow.get_flow_from_file("tests/testdata/test_flow_2.yaml")
+            flow = self.flow.get_flow_from_file("tests/flows/testdata/test_flow_2.yaml")
         block = flow[0]["block_type"](**flow[0]["block_config"])
 
         self.assertEqual(block.block_config["introduction"], "intro")
