--- tests/testdata/testdata.py
@@ -1,25 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-documents = [
-    """Knowledge is an awareness of facts, a familiarity with individuals and situations, 
-      or a practical skill. Knowledge of facts, also called propositional knowledge, is often characterized 
-      as true belief that is distinct from opinion or guesswork by virtue of justification. 
-      While there is wide agreement among philosophers that propositional knowledge is a form of true belief, 
-      many controversies focus on justification. This includes questions like how to understand justification, 
-      whether it is needed at all, and whether something else besides it is needed. 
-      These controversies intensified in the latter half of the 20th century due to a series of thought experiments 
-      called Gettier cases that provoked alternative definitions."""
-]
-
-long_line_documents = [
-    (
-        "Knowledge is an awareness of facts, a familiarity with individuals and situations,"
-        "or a practical skill. Knowledge of facts, also called propositional knowledge, is often characterized "
-        "as true belief that is distinct from opinion or guesswork by virtue of justification. "
-        "While there is wide agreement among philosophers that propositional knowledge is a form of true belief, "
-        "many controversies focus on justification. This includes questions like how to understand justification, "
-        "whether it is needed at all, and whether something else besides it is needed. "
-        "These controversies intensified in the latter half of the 20th century due to a series of thought experiments "
-        "called Gettier cases that provoked alternative definitions."
-    )
-]
